pyproject.toml
[tool.poetry]
name = "planchet-bot"
version = "0.1.0"
description = ""
authors = ["Kintaro Oe <tarokinoe@gmail.com>"]

[tool.poetry.scripts]
manage = 'planchet_bot.manage:main'

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.10"
typer = {extras = ["all"], version = "^0.4.0"}
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
